Split `FollowButton` into presentational and connected parts
Extract the API and state handling into a new `CrateFollowButton` and reduce
`FollowButton` to a presentational component driven by props. The visual
state is expressed as a single `status` of `loading`, `following`, or
`not-following`, so the loading and follow states can not contradict
each other. `CrateHeader` now renders `CrateFollowButton`.
